Completion of ImageAggregator utility
ImageAggregator is a reimplementation of code given to me by Sam
Ruggieri (owner of Voyager Games, the client) which constructed a larger
picture from individual sprite pictures.

In this commit I implemented his code in my program, going line by line
through it not only so I could tailor it to fit my program, but also to
understand its workings in full.

Some unforeseen bugs were found and fixed.

Bug #1 - When the program was used multiple times in the same directory
the previous Output.png image was used in successive regenerations of
Output.png. This occurred because of the manner in which file paths are
loaded into memory. A check was inserted to make sure files named
Output.png is ignored.

Bug #2 - Only discovered after the first bug was fixed, I found that the
program would only generate proper output files when the given Width and
Height were equal. After troubleshooting what pixels were being written
to what coordinates and to which files I discovered the problem: A
calculation for a proper X-coordinate value was accidentally doing the
calculation for the Y-coordinate, and vice versa (i.e. calculations were
% and /, but needed to be / and %). Should've seen this sooner. I fought
this sort of problem constantly in my previous HillClimber program.

Bugs are fixed, and ImageAggregator is now functioning 100% as intended.

Next to come is the GIF-making utility class!