Work complete!
A massive re-working of the ImageAggregator class was done to break down
the code into smaller, publically-accessible methods.

A special thanks goes to Elliot Kroo whose GifSequenceWriter.java Class
I found on Java. This component does the heavy lifting for making the
GIF. I merely made a wrapper of sorts that adapted it to my program.

The project is complete. The VG-TFR-TopFolder\ folder is an example of
what my program had to work with. That folder contains a bunch of
numbered folders within, and each folder contains a single frame of a
game sprite's animation. My program stitches those together into a GIF.